Dev 3

+++ app/src/main/java/com/zellycookies/pineapple/utils/DeleteAccountActivity.kt
@@ -1,4 +1,4 @@
-package com.zellycookies.pineapple
+package com.zellycookies.pineapple.utils
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
@@ -6,13 +6,20 @@ import android.os.Bundle
 import android.util.Log
 import android.widget.Button
 import android.widget.EditText
+import android.widget.Toast
 import com.google.firebase.auth.AuthCredential
 import com.google.firebase.auth.EmailAuthProvider
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
+import com.zellycookies.pineapple.R
 import com.zellycookies.pineapple.introduction.IntroductionMain
 
 class DeleteAccountActivity : AppCompatActivity() {
+    //firebase
+    private var mAuth: FirebaseAuth? = null
+    private var mAuthListener: FirebaseAuth.AuthStateListener? = null
+    private var user : FirebaseUser? = null
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_delete_account)
@@ -24,25 +31,30 @@ class DeleteAccountActivity : AppCompatActivity() {
     }
 
     // Delete account
-    fun deleteAccount() {
-        val inputEmail : EditText = findViewById(R.id.input_email)
+    private fun deleteAccount() {
         val inputPassword : EditText = findViewById(R.id.input_password)
 
         Log.d(TAG, "Getting users Credentials")
-        var user : FirebaseUser? = FirebaseAuth.getInstance().currentUser
         val credential : AuthCredential = EmailAuthProvider.getCredential(
-            inputEmail.text.toString(), inputPassword.text.toString()
-        )
+            user!!.email!!, inputPassword.text.toString())
+        Log.d(TAG, user!!.toString())
         Log.d(TAG, "Deleting user's account")
-        user!!.reauthenticate(credential).addOnCompleteListener {
-            Log.d(TAG, "User re-authenticated")
-            user!!.delete().addOnCompleteListener { task ->
-                if (task.isSuccessful) {
-                    FirebaseAuth.getInstance().signOut()
-                    user = null
-                    checkCurrentUser(user)
-                    Log.d(TAG, "User account deleted")
+        user!!.reauthenticate(credential).addOnCompleteListener { taskAuth ->
+            if (taskAuth.isSuccessful) {
+                Log.d(TAG, "User re-authenticated")
+                user!!.delete().addOnCompleteListener { task ->
+                    if (task.isSuccessful) {
+                        mAuth!!.signOut()
+                        finish()
+                    }
                 }
+            } else {
+                Toast.makeText(
+                    this@DeleteAccountActivity,
+                    "Failed to re-authenticate",
+                    Toast.LENGTH_SHORT
+                ).show()
+                Log.d(TAG, "Failed to re-authenticate user")
             }
         }
     }
@@ -57,8 +69,8 @@ class DeleteAccountActivity : AppCompatActivity() {
 
     private fun setupFirebaseAuth() {
         Log.d(TAG, "setupFirebaseAuth: check user")
-        var mAuth = FirebaseAuth.getInstance()
-        val mAuthListener = FirebaseAuth.AuthStateListener { firebaseAuth ->
+        mAuth = FirebaseAuth.getInstance()
+        mAuthListener = FirebaseAuth.AuthStateListener { firebaseAuth ->
             val user = firebaseAuth.currentUser
 
             //check if the user is logged in
@@ -71,6 +83,7 @@ class DeleteAccountActivity : AppCompatActivity() {
                 Log.d(TAG, "onAuthStateChanged: signed_out")
             }
         }
+        user = FirebaseAuth.getInstance().currentUser
     }
 
     companion object {
